Fix remaining MSVC warnings in our code
- FrameRenderer.hpp: add explicit casts for int to uint8_t conversions
- Saa5050.hpp: change m_last_graphics_data from uint16_t to uint8_t
- SidewaysService.hpp: cast proto slot numbers to uint8_t after validation
- Platform.hpp: add cross-platform get_env() using _dupenv_s on Windows
- RomPaths.hpp: use safe get_env() helper instead of std::getenv
- ServerMain.hpp: use platform::get_env() instead of std::getenv

+++ src/core/include/beebium/Saa5050.hpp
@@ -558,8 +558,8 @@ private:
     TeletextCharset m_charset = TeletextCharset::Alpha;
     TeletextCharset m_graphics_charset = TeletextCharset::ContiguousGraphics;
 
-    // Graphics hold state (12-bit expanded font data)
-    uint16_t m_last_graphics_data = 0;
+    // Graphics hold state
+    uint8_t m_last_graphics_data = 0;
 
     // Double height state
     uint8_t m_raster_shift = 0;
